Added LabelBelowBox and AlternativeLabelling settings
Small contribution

LabelBelowBox setting: when True, it will place each text label below the object's box. Default: False.

This is because in my case, my cameras usually detect persons tending to be walking on top of the image, and labels where being cut-off, out of frame.

AlternativeLabelling setting: When True, and also DrawMode is set to "Match" then  the label will show just a reference number, which the will be correlated while text listing each object on Telegram or Pushbullet notification services. Default: False.

This is because in my case, when more than one person is detected walking in groups, labels ended on top of each other, illegible.