streamlining snapshot processing, adding web report
Lots of changes:

Now the snapshot is always processed ("dressed") with its boundaries / graphics BEFORE reaching the sending (if any is configured) the notifications. The rationale behind this: The user may be interested in just getting the local copy of the snapshot and not really interested on being notified in any way. This new behavior also simplified the code and streamlined its workflow.

(for future upgrade: The user may be interested even in just storing just the parts or boundaries with the objects of interest, instead of the whole frame, for further cataloging / processing)

Added a new way for file-naming each snapshot. This is active only if the AlternativeLabelling flag is true. This new filenaming includes the following changes:

1. Gets rid of the camera name at the start of the filename, since snapshots are saved inside a folder which is already named after that camera.

2. Gets rid of the "frame" time (after the seconds) since it actually doesn´t add any useful information.

3. Adds a new suffix after the "timestamp" string, separated by "-" following by the number of objects detected (this will be used later, for the new web / graph feature).

4. Shortens the filename extension, from .jpeg to .jpg

Added an experimental web page, including a report on how many images (and objects inside those images) SynoAI has managed over the last 24 hours of activity, along with a couple of other interesting data.

It might be the stepping stone into adding lots of functionality in the future:

1. We may add an "online" web browsing for the images, based on the user clicking in any bar from the graph

2. A selector for "date" which allows to investigate older dates and its images

3. Try to get more interesting data out of the available information: I.E: further analyze movement / people / objects grouping

4. Add a web based configuration page for SynoAI config.