Improved Y axis labelling on web graph
Now, at start time, if the number of snapshots / objects is low, it will only reference the top (MAX) value. It will start detailing a scale of values if the detected snapshots / objects is big enough to warrant a minimal precision as reference in the number stamped at each step in the Y axis.