Added the Gallery webpage
Now the SynoAI webserver allows you to see a graph depicting the last 24 hours of activity on your cameras.

Clicking in any hour bar will zoom into a graph  showing the minutes which had activity for that camera.

(On this commit): Clicking on a minute bar will open up a gallery showing the thumbnails from the available snapshots taken inside that minute. Depending the quantity, it will resize each snapshot in order to better accommodate them.

Clicking in one of the snapshot's thumbnails, it will open the original full sized snapshot.